+++ src/backend/core/api/viewsets.py
@@ -4425,6 +4425,7 @@ class ConfigView(drf.views.APIView):
 
         array_settings = [
             "MESSAGES_PUBLIC_URL",
+            "PROJECTS_PUBLIC_URL",
             "STORAGE_UNIFIED_ENABLED",
             "DOCS_DRIVE_ENABLED",
             "DOCS_PUBLIC_URL",